fix features processing
